[refactor] move the cluster objects into the level 0 directory of the package
