+++ vcontainer/services/coe/v2/cluster/iresponses.go
@@ -2,13 +2,12 @@ package cluster
 
 import (
 	"github.com/vngcloud/vcontainer-sdk/vcontainer/objects"
-	"github.com/vngcloud/vcontainer-sdk/vcontainer/services/coe/v2/cluster/obj"
 )
 
 // ******************************************** Response of GetCluster API *********************************************
 
 type IGetResponse interface {
-	ToClusterObject() *obj.Cluster
+	ToClusterObject() *objects.Cluster
 }
 
 // *************************************** Response of Update Secgroup rule API ****************************************
